requirements.txt
customtkinter>=5.2.0
anthropic>=0.39.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
markdown>=3.5.0
